implemented flip and rotation, need to tweak rotation

+++ src/types/snake.rs
@@ -46,10 +46,10 @@ impl Snake {
         //draw head
         // Self::draw_snake_part(grid, &self.head, DARKGREEN);
         let rotation = match self.dir.0 {
-            -1 => -90,
-            0 => 0,
-            1 => 90,
-            _ => 0,
+            -1 => -90.0,
+            0 => 0.0,
+            1 => 90.0,
+            _ => 0.0,
         };
         draw_texture_ex(
             &self.head_texture,
@@ -58,7 +58,7 @@ impl Snake {
             WHITE,
             DrawTextureParams {
                 dest_size: Some(vec2(grid.square_size, grid.square_size)),
-                rotation: rotation as f32,
+                rotation: rotation,
                 flip_y: self.dir.1 > 0,
                 ..Default::default()
             },
